Split different models to separated files
