+++ feed/models/__init__.py
@@ -1,0 +1,2 @@
+from .post import Post, PostManager
+from .comment import Comment
