+++ feed/models/post.py
@@ -1,11 +1,7 @@
 from django.db import models
 from django.utils import timezone
 from django.contrib.auth.models import User
-
-
-class PostManager(models.Manager):
-    def main_feed(self):
-        return self.order_by('-date_posted')
+from .postManager import PostManager
 
 
 class Post(models.Model):
@@ -19,12 +15,3 @@ class Post(models.Model):
 
     def __str__(self):
         return self.title
-
-
-class Comment(models.Model):
-    author = models.ForeignKey(User, on_delete=models.CASCADE)
-    comment_parent = models.ForeignKey('self', on_delete=models.CASCADE, blank=True, null=True)
-    post_parent = models.ForeignKey(Post, related_name="comments", on_delete=models.CASCADE)
-    date_posted = models.DateTimeField(default=timezone.now)
-    content = models.TextField()
-    comments = models.Manager()
